% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{chance_to_draw_n_games}
\alias{chance_to_draw_n_games}
\title{Calculates the chance to draw out of n matches}
\usage{
chance_to_draw_n_games(p, n)
}
\arguments{
\item{p}{probability of first (or second) player winning match}

\item{n}{number of matches}
}
\value{
The decimal chance for a draw
}
\description{
Calculates the chance to draw out of n matches
}
\examples{
chance_to_draw_n_games(0.4, 4) # Draw chance if one player has p=0.4 in four matches
}
