% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{chance_to_win_n_games}
\alias{chance_to_win_n_games}
\title{Calculate win chance after multiple matches}
\usage{
chance_to_win_n_games(p, n)
}
\arguments{
\item{p}{probability for player to win a single match}

\item{n}{number of total matches playes}
}
\value{
The decimal chance of winning a game
}
\description{
Chance of a player winning the majority of n matches. Draws count not as a
win
}
\examples{
chance_to_win_n_games(0.55,5) # Chance for player with p=0.55 to win best of 5 matches
}
