% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{kelly_back_dec}
\alias{kelly_back_dec}
\title{Kelly for back bet}
\usage{
kelly_back_dec(price, p, commision_rate)
}
\arguments{
\item{price}{Price to back in decimal odds}

\item{p}{Probability of event to to materialise}

\item{commision_rate}{Rate of commision charged on WINNINGS}
}
\value{
Kelly optimised fraction of stake relative to bank
}
\description{
Kelly for back bet
}
\examples{
kelly_back_dec(2,0.5,0.05)
}
