% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{kelly_criterion}
\alias{kelly_criterion}
\title{The Kelly criterion}
\usage{
kelly_criterion(p, alpha_w, alpha_l)
}
\arguments{
\item{p}{The objective probability of the event}

\item{alpha_w}{The return multiplier in case of the event happening}

\item{alpha_l}{The return multiplier in case of the event not happening}
}
\value{
The Kelly optimised fraction of the bankroll that should be bet
}
\description{
The Kelly criterion
}
\examples{
kelly_criterion(0.5,1,1)

}
\references{
Thorp, Edward O. (1997; revised 1998). The Kelly Criterion in Blackjack, Sports Betting, and the Stock Market. \url{http://www.eecs.harvard.edu/cs286r/courses/fall12/papers/Thorpe_KellyCriterion2007.pdf}
}
