% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{kelly_lay_dec}
\alias{kelly_lay_dec}
\title{Kelly for lay bet}
\usage{
kelly_lay_dec(price, p, commision_rate)
}
\arguments{
\item{price}{Price at which to lay}

\item{p}{Base probability of event that is being laid}

\item{commision_rate}{Rate of commision charged on WINNINGS}
}
\value{
Kelly optimised fraction of stake relative to bank
}
\description{
Kelly for lay bet
}
